uitgever-service/requirements.txt
pika==1.2.0
python-dotenv==1.0.0